Move some files from knative-operator/.../common to .../common/serving
