+++ knative-operator/pkg/common/serving/openshift.go
@@ -1,4 +1,4 @@
-package common
+package serving
 
 import (
 	"context"
